% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-generic-models.R
\name{a_model}
\alias{a_model}
\title{Function for defining a cgeneric model for a.}
\usage{
a_model(y0, dist_to_s0, init, priors, is_fixed = rep(FALSE, 2), debug = FALSE)
}
\description{
The input variables are:
y0: A vector of equal length to the data (y_inla) used for running R-INLA,
such that y0\link{i} contains the value at the conditioning site for the same
time as when y_inla\link{i} was observed.
dist_to_s0: A vector of equal length to y_inla, such that dist_to_s0\link{i} contains
the distance from the location where y_inla\link{i} was observed, to the conditioning
site where y0\link{i} was observed.
init: Initial values for the model parameters.
priors: Priors for the model parameters. This is a list with one element per model parameter,
with names "lambda" and "kappa". Each element contains hyperparameters
for the prior of that specific model parameter.
is_fixed: A vector of bools, stating whether each model parameter should be fixed to its
initial value or if it should be estimated.
debug: A boolean stating if R-INLA should print debug information or not.
}
