% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{aggregated_ecdf}
\alias{aggregated_ecdf}
\title{Compute the aggregated empirical cumulative distribution function for all
data inside a circle with a given radius. Input variables are:
data: An (n x d)-dimensional matrix of observations.
coords: A (d x 2)-dimensional matrix with the coordinates of the data.
center: A 2-dimensional vector containing the coordinates of the center we use
for computing the aggregated ECDF.
radius: The radius of the circle used for computing the aggregated ECDF.}
\usage{
aggregated_ecdf(data, coords, center, radius, zero_threshold = 0)
}
\description{
Compute the aggregated empirical cumulative distribution function for all
data inside a circle with a given radius. Input variables are:
data: An (n x d)-dimensional matrix of observations.
coords: A (d x 2)-dimensional matrix with the coordinates of the data.
center: A 2-dimensional vector containing the coordinates of the center we use
for computing the aggregated ECDF.
radius: The radius of the circle used for computing the aggregated ECDF.
}
