% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional-likelihood.R
\name{dconditional_r}
\alias{dconditional_r}
\title{Compute the (log-)likelihood of the conditional extremes distribution when
a and b are given. We assume that a has already been subtracted from x,
i.e. x = (y - a) for some observations y.
This function is mainly written to evaluate the speed and correctness of the Rcpp
function dconditional()}
\usage{
dconditional_r(x, A, B, sigma0, nugget, logd = TRUE, na_rm = TRUE)
}
\description{
The input variables are:
x: an (n x d)-dimensional matrix of observations where a has already been subtracted.
A: The projection matrix used for building the SPDE approximation
B: An (m x n)-dimensional matrix containing the values of b at the m triangular mesh nodes,
for each of the n threshold exceedances at the conditioning sites of interest
sigma0: The covariance matrix of the m Gaussian random variables in the triangular mesh.
nugget: The variance of the nugget effect.
logd: Boolean explaining if we should return the log-likelihood or the likelihood.
na_rm: Boolean explaining how we should treat NA values. If na_rm = false, then
any column of x that returns an NA value will result in an NA value in the output.
If na_rm = true, then we remove the NA values before computing the likelihood.
So if e.g. a column has 3 NA variables, then we remove these, and then we compute
the likelihood for a (d-3)-dimensional Gaussian random variable.
}
