% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_euclid}
\alias{dist_euclid}
\title{Compute the Euclidean distance between x and y, where
x is either a row vector, or a matrix of row vectors,
and y is a matrix of row vectors}
\usage{
dist_euclid(x, y)
}
\description{
Compute the Euclidean distance between x and y, where
x is either a row vector, or a matrix of row vectors,
and y is a matrix of row vectors
}
