% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{empirical_moments}
\alias{empirical_moments}
\title{Compute the empirical estimators of the first two conditional moments of a dataset
for a set of threshold exceedances and distances.
Function input:
\itemize{
\item data is a list of data on the same format as that created by
the function extract_extreme_fields().
\item unique_dist is an integer vector of distances d for which we
should estimate the conditional moments μ(d; y_0) and ζ(d; y_0)
\item unique_y0 is a vector of threshold exceedances y0 for which we
should estimate the conditional moments μ(d; y_0) and ζ(d; y_0)
}}
\usage{
empirical_moments(data, unique_dist, unique_y0)
}
\description{
Compute the empirical estimators of the first two conditional moments of a dataset
for a set of threshold exceedances and distances.
Function input:
\itemize{
\item data is a list of data on the same format as that created by
the function extract_extreme_fields().
\item unique_dist is an integer vector of distances d for which we
should estimate the conditional moments μ(d; y_0) and ζ(d; y_0)
\item unique_y0 is a vector of threshold exceedances y0 for which we
should estimate the conditional moments μ(d; y_0) and ζ(d; y_0)
}
}
