% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{execute_shell_script}
\alias{execute_shell_script}
\title{Execute a shell script, and provide feedback if it crashes}
\usage{
execute_shell_script(command, args = character(), ...)
}
\description{
Execute a shell script, and provide feedback if it crashes
}
