% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{extract_extreme_fields}
\alias{extract_extreme_fields}
\title{Search through data after threshold exceedances at a set of conditioning sites.
Extract relevant information and observations for each of the threshold exceedances.}
\usage{
extract_extreme_fields(
  data,
  coords,
  s0_index,
  threshold,
  n,
  r,
  verbose = FALSE,
  n_cores = 1
)
}
\description{
The input variables are:
data: An (n x d)-dimensional matrix of observations.
coords: A (d x 2)-dimensional matrix with the coordinates of the data.
s0_index: a vector of indices describing which of the d coordinates are chosen
as conditioning sites.
threshold: The threshold t for the conditional extremes model.
n, r: Used for only extracting a subset of the observations in a circle around
the conditioning site. See the docs for extract_thinned_out_circles() for more info.
}
\details{
The output of the function is a list where each element is a list or vector
of length equal, containing these variables:
s0_index: This is a subset of all the s0_indices given as input such that each
conditioning site has at least one threshold exceedance.
s0: A list where element nr. i is a 2-dimensional vector with the coordinates of
conditioning site nr. i.
obs_index: A list where element nr. i is a vector of indices for all the
locations in coords we use for performing inference using conditioning site nr. i.
dist_to_s0: A list where element nr. i is a vector with all the distances between
conditioning site nr. i and the locations used for inference with conditioning
site nr. i.
y0: A list where element nr. i is a vector of all threshold exceedances for
conditioning site nr. i.
y: A list where element nr. i is a matrix consisting of one column per threshold
exceedance in y0[\link{i}]. Each column contains observations from the locations
described in obs_index[\link{i}].
time_index: A list where element nr. i contains the times of all threshold
exceedances from conditioning site nr. i.
n: A vector where element nr. i is the number of threshold exceedances at
conditioning site nr. i.
}
