% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{extract_thinned_out_circles}
\alias{extract_thinned_out_circles}
\title{Given a matrix of coordinates on a regular grid, extract a subset of the coordinates
that consists of circles centred around some center, with varying degrees of
densities. The input variables are:
coords: A matrix with two columns and multiple rows, containing (x, y) coordinates
in a regular grid and with a distance-preserving projection.
center: One (x, y) tuple of coordinates, explaining the center of the circles we
wish to extract.
n: A vector of densities for the different circles we wish to extract. If the ith
value of n is k, then we only keep every kth of the original coords, both in
the x-direction and in the y-direction.
r: A vector of radii for the different circles we wish to extract. If e.g.
n = (1, 3) and r = (2, 5), then we will extract every coordinate that is closer
to the center than a radius of 2. Then, we will only keep every 3rd coordinate,
both in x-direction and y-direction, for all coordinates that have a distance
between 2 and 5 from the center. Finally, all coords that are further away from
the center than 5 will be dropped.
index_only: A bool describing if the function should return the extracted
coordinates, or only their indices inside the coords object.}
\usage{
extract_thinned_out_circles(coords, center, n, r = Inf, index_only = FALSE)
}
\description{
Given a matrix of coordinates on a regular grid, extract a subset of the coordinates
that consists of circles centred around some center, with varying degrees of
densities. The input variables are:
coords: A matrix with two columns and multiple rows, containing (x, y) coordinates
in a regular grid and with a distance-preserving projection.
center: One (x, y) tuple of coordinates, explaining the center of the circles we
wish to extract.
n: A vector of densities for the different circles we wish to extract. If the ith
value of n is k, then we only keep every kth of the original coords, both in
the x-direction and in the y-direction.
r: A vector of radii for the different circles we wish to extract. If e.g.
n = (1, 3) and r = (2, 5), then we will extract every coordinate that is closer
to the center than a radius of 2. Then, we will only keep every 3rd coordinate,
both in x-direction and y-direction, for all coordinates that have a distance
between 2 and 5 from the center. Finally, all coords that are further away from
the center than 5 will be dropped.
index_only: A bool describing if the function should return the extracted
coordinates, or only their indices inside the coords object.
}
