% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-evaluation.R
\name{extremal_correlation}
\alias{extremal_correlation}
\title{Calculate chi and chibar at the quantiles u for two random variables
This one is heavily based on the implementation in
https://github.com/harrysouthworth/texmex/blob/master/R/chi.R}
\usage{
extremal_correlation(p, x, y, truncate = TRUE)
}
\description{
Calculate chi and chibar at the quantiles u for two random variables
This one is heavily based on the implementation in
https://github.com/harrysouthworth/texmex/blob/master/R/chi.R
}
