% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-evaluation.R
\name{extremal_correlation_in_dist_groups}
\alias{extremal_correlation_in_dist_groups}
\title{Estimate the extremal correlation where we group together all observations
that have similar distances to s0
y0: n-dimensional vector of y(s0)
y: (nxd)-dimensional matrix of observations with Laplace marginals
dist_to_s0: d-dimensional vector of lenghts
dist_groups: vector of distances such that group nr. i contains all locations with
dist_groups\link{i} <= dist_to_s0 < dist_groups\link{i + 1}
p: vector of probabilities for computing χ}
\usage{
extremal_correlation_in_dist_groups(y, y0, dist_to_s0, dist_groups, p)
}
\description{
Estimate the extremal correlation where we group together all observations
that have similar distances to s0
y0: n-dimensional vector of y(s0)
y: (nxd)-dimensional matrix of observations with Laplace marginals
dist_to_s0: d-dimensional vector of lenghts
dist_groups: vector of distances such that group nr. i contains all locations with
dist_groups\link{i} <= dist_to_s0 < dist_groups\link{i + 1}
p: vector of probabilities for computing χ
}
