% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-c-matrix.R
\name{get_C}
\alias{get_C}
\title{Compute the matrix C that solves the equation
C H^-1 C^T = H^-1 J H^-1}
\usage{
get_C(H, J)
}
\description{
Compute the matrix C that solves the equation
C H^-1 C^T = H^-1 J H^-1
}
