% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{get_radar_url}
\alias{get_radar_url}
\title{For a given date on a format that can be understood by lubridate,
return a vector of urls containing all the radar data archives
for that month}
\usage{
get_radar_url(date)
}
\description{
For a given date on a format that can be understood by lubridate,
return a vector of urls containing all the radar data archives
for that month
}
