% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{get_s0_index}
\alias{get_s0_index}
\title{Given a matrix of coordinates, return a sub-grid of the coordinates
with resolution (delta_s0 x delta_s0)}
\usage{
get_s0_index(coords, delta_s0 = 1)
}
\description{
Given a matrix of coordinates, return a sub-grid of the coordinates
with resolution (delta_s0 x delta_s0)
}
