% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-c-matrix.R
\name{h_loggamma}
\alias{h_loggamma}
\title{The negative Hessian of the log of a log-gamma pdf.
If tau ~ gamma(a, b), then x = log(tau) ~ loggamma(a, b)}
\usage{
h_loggamma(x, rate)
}
\description{
The negative Hessian of the log of a log-gamma pdf.
If tau ~ gamma(a, b), then x = log(tau) ~ loggamma(a, b)
}
