% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-c-matrix.R
\name{h_PC}
\alias{h_PC}
\title{The negative Hessian of the log of a Gaussian Matérn PC prior,
where theta = (log(ρ), log(σ))}
\usage{
h_PC(theta, rho_0, sigma_0, p_rho, p_sigma)
}
\description{
The negative Hessian of the log of a Gaussian Matérn PC prior,
where theta = (log(ρ), log(σ))
}
