% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{keef_sampling}
\alias{keef_sampling}
\title{Sample from the global conditional extremes distribution of \link{x | max(x) > t},
using the simulation algorithm of Keef et al. (2013).}
\usage{
keef_sampling(
  n,
  a_func,
  b_func,
  Q,
  tau,
  threshold,
  dist_to_s0,
  dist_to_s0_from_mesh,
  A,
  verbose = FALSE
)
}
\description{
The arguments of the function are:
n: Number of samples.
a_func: Function a(y0, dist_to_s0) that takes in two vectors and returns a
matrix where element (i, j) is the value of a(y0\link{j}, dist_to_s0\link{i}).
b_func: Function b(y0, dist_to_s0) that takes in two vectors and returns a
matrix where element (i, j) is the value of b(y0\link{j}, dist_to_s0\link{i}).
b_func is different from a_func in that a_func requires the distance to s0
for all locations with observations, while b_func requires the distance to s0
for all locations on the triangular mesh used for defining the SPDE.
Q: Precision matrix of the weights used for building the SPDE approximation.
tau: Precision of the nugget effect.
threshold: The threshold t used for defining the conditional extremes distribution.
dist_to_s0: List containing the distances to s0 for all observation locations,
used for computing a. The list contains one vector of distances for each
possible conditioning site in the domain of interest. Each of the distance
vectors must have the same length.
dist_to_s0_from_mesh: List containing the distances to s0 for all locations in
the mesh, used for computing b. The list contains one vector of distances for each
possible conditioning site in the domain of interest. Each of the distance
vectors must have the same length.
A: List of projection matrices used for building the SPDE approximation. The list
contains one projection matrix for each possible conditioning site in the
domain of interest. Each projection matrix must contain the same number of
rows (equal to the lengths of dist_to_s0) and columns (equal to the lengths of
dist_to_s0_from_mesh).
verbose: bool that states if the function should display progress or not.
}
