% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Compute the logarithm of the sum of the exponents of a vector x.
This function is usable in settings where components x_i of x are so large/big that
exp(x_i) = Inf, or exp(x_i) = 0}
\usage{
log_sum_exp(x, na.rm = FALSE)
}
\description{
Compute the logarithm of the sum of the exponents of a vector x.
This function is usable in settings where components x_i of x are so large/big that
exp(x_i) = Inf, or exp(x_i) = 0
}
