% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional-likelihood.R
\name{loglik_conditional}
\alias{loglik_conditional}
\title{Comput the composite log-likelihood for the global conditional extremes model.}
\usage{
loglik_conditional(
  y,
  y0,
  a_func,
  b_func,
  Q,
  tau,
  dist_to_s0,
  dist_to_s0_from_mesh,
  A,
  n_cores = 1,
  na.rm = TRUE,
  use_r = FALSE
)
}
\description{
The input variables are:
y: A list of matrices. Matrix nr. i has dimension (d_i x n_i) and contains n_i
d_i-dimensional vectors of observations for conditioning site nr. i.
y0: A list of vectors. Vector nr. i has dimension n_i and contains the threshold
exceedances from conditioning site nr. i.
a_func: A function a(y0, dist_to_s0) that takes in two vectors and returns a
matrix where element (i, j) is the value of a(y0\link{j}), dist_to_s0\link{i}).
b_func: A function b(y0, dist_to_s0) that takes in two vectors and returns a
matrix where element (i, j) is the value of b(y0\link{j}, dist_to_s0\link{i}).
b_func is different from a_func in that a_func requires the distance to s0
for all locations with observations, while b_func requires the distance to s0
for all locations on the triangular mesh used for defining the SPDE.
Q: Either a precision matrix, or a list of precision matrices, of the mesh for
each conditioning site. The list must have length 1 or length(y0)
sigma: The covariance matrix of the weights used for building the SPDE approximation.
tau: The precision of the nugget effect.
dist_to_s0: List containing the distances to s0 for all observation locations,
used for computing a. The list contains one vector of distances for each
conditioning site. Vector nr. i has dimension d_i.
dist_to_s0_from_mesh: List containing the distances to s0 for all locations in
the mesh, used for computing b. The list contains one vector of distances for each
conditioning site.
A: List of projection matrices used for building the SPDE approximation. The list
contains one projection matrix for each conditioning site.
Each projection matrix must contain the same number of columns, equal to the
lengths of dist_to_s0_from_mesh. Projection matrix nr. i has d_i rows.
n_cores: Number of cores to use for computing the log-likelihood in parallel. We
cannot use more cores than the number of conditioning sites.
na.rm: Should NA's be removed or not? If na.rm = FALSE, then
any column of x that returns an NA value will result in an NA value in the output.
If na.rm = TRUE, then we remove the NA values before computing the likelihood.
So if e.g. a column has 3 NA variables, then we remove these, and then we compute
the likelihood for a (d-3)-dimensional Gaussian random variable.
use_r: A boolean, mostly used for testing the code. Should we perform the actual computations
using R code or using Rcpp code?
}
