% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_cgeneric}
\alias{make_cgeneric}
\title{Call a makefile used for compiling and linking cgeneric scripts,
in order to use them with R-INLA}
\usage{
make_cgeneric(cmd)
}
\description{
Call a makefile used for compiling and linking cgeneric scripts,
in order to use them with R-INLA
}
