% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-distributions.R
\name{marginal_distribution}
\alias{marginal_distribution}
\title{Compute the empirical marginal distribution of the vector x and return
a function F(x) that is equal to the marginal distribution when x > zero_val
and returns NA when x <= zero_val.}
\usage{
marginal_distribution(x, zero_threshold = 0)
}
\description{
Compute the empirical marginal distribution of the vector x and return
a function F(x) that is equal to the marginal distribution when x > zero_val
and returns NA when x <= zero_val.
}
