% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{plot_chi}
\alias{plot_chi}
\title{Plot the estimates of χ_p(d), created using the empirical_chi() function}
\usage{
plot_chi(chi)
}
\description{
Plot the estimates of χ_p(d), created using the empirical_chi() function
}
