% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progress_bar}
\alias{progress_bar}
\title{Create a progress bar for tracking long-running processes.
This is a thin wrapper around the progress package}
\usage{
progress_bar(n)
}
\description{
Create a progress bar for tracking long-running processes.
This is a thin wrapper around the progress package
}
