% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{rconditional}
\alias{rconditional}
\title{Function for simulating from the conditional extremes model given the values of y(s0).
The arguments of the function are:
y0: A list of vectors containing the values of y(s0) for each of the possible
conditioning sites s0. The vectors can have length 0.
a_func: Function a(y0, dist_to_s0) that takes in two vectors and returns a
matrix where element (i, j) is the value of a(y0\link{j}, dist_to_s0\link{i}).
b_func: Function b(y0, dist_to_s0) that takes in two vectors and returns a
matrix where element (i, j) is the value of a(y0\link{j}, dist_to_s0\link{i}).
b_func is different from a_func in that a_func requires the distance to s0
for all locations with observations, while b_func requires the distance to s0
for all locations on the triangular mesh used for defining the SPDE.
Q: Precision matrix of the weights used for building the SPDE approximation.
tau: Precision of the nugget effect.
threshold: The threshold t used for defining the conditional extremes distribution.
dist_to_s0: List containing the distances to s0 for all observation locations,
used for computing a. The list contains one vector of distances for each
possible conditioning site in the domain of interest. Each of the distance
vectors may be of different size.
dist_to_s0_from_mesh: List containing the distances to s0 for all locations in
the mesh, used for computing b. The list contains one vector of distances for each
possible conditioning site in the domain of interest. Each of the distance
vectors must be of the same size.
A: List of projection matrices used for building the SPDE approximation. The list
contains one projection matrix for each possible conditioning site in the
domain of interest. Each projection matrix may contain a different number of
rows, but all must have the same number of columns, which is equal to the number
nodes in the triangular mesh.}
\usage{
rconditional(y0, a_func, b_func, Q, tau, dist_to_s0, dist_to_s0_from_mesh, A)
}
\description{
Function for simulating from the conditional extremes model given the values of y(s0).
The arguments of the function are:
y0: A list of vectors containing the values of y(s0) for each of the possible
conditioning sites s0. The vectors can have length 0.
a_func: Function a(y0, dist_to_s0) that takes in two vectors and returns a
matrix where element (i, j) is the value of a(y0\link{j}, dist_to_s0\link{i}).
b_func: Function b(y0, dist_to_s0) that takes in two vectors and returns a
matrix where element (i, j) is the value of a(y0\link{j}, dist_to_s0\link{i}).
b_func is different from a_func in that a_func requires the distance to s0
for all locations with observations, while b_func requires the distance to s0
for all locations on the triangular mesh used for defining the SPDE.
Q: Precision matrix of the weights used for building the SPDE approximation.
tau: Precision of the nugget effect.
threshold: The threshold t used for defining the conditional extremes distribution.
dist_to_s0: List containing the distances to s0 for all observation locations,
used for computing a. The list contains one vector of distances for each
possible conditioning site in the domain of interest. Each of the distance
vectors may be of different size.
dist_to_s0_from_mesh: List containing the distances to s0 for all locations in
the mesh, used for computing b. The list contains one vector of distances for each
possible conditioning site in the domain of interest. Each of the distance
vectors must be of the same size.
A: List of projection matrices used for building the SPDE approximation. The list
contains one projection matrix for each possible conditioning site in the
domain of interest. Each projection matrix may contain a different number of
rows, but all must have the same number of columns, which is equal to the number
nodes in the triangular mesh.
}
