% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{rnorm_spde}
\alias{rnorm_spde}
\title{This is a wrapper for the inla.qsample() function for sampling from a Gaussian
random field with sparse precision matrix Q. The wrapper allows us to get
reproduceable results if the seed outside the function scope is known.
It also suppresses warnings and removes some unnecessary column and row names}
\usage{
rnorm_spde(n, Q, ...)
}
\description{
This is a wrapper for the inla.qsample() function for sampling from a Gaussian
random field with sparse precision matrix Q. The wrapper allows us to get
reproduceable results if the seed outside the function scope is known.
It also suppresses warnings and removes some unnecessary column and row names
}
