% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-generic-models.R
\name{spatial_conditional_extremes_cgeneric_model}
\alias{spatial_conditional_extremes_cgeneric_model}
\title{Function for defining a cgeneric model for the entire spatial conditional extremes model.}
\usage{
spatial_conditional_extremes_cgeneric_model(
  spde,
  y0,
  n,
  init,
  priors,
  dist_to_s0,
  dist_to_s0_from_mesh,
  C = NULL,
  theta_star = NULL,
  debug = FALSE
)
}
\description{
The input variables are:
spde: An inla.spde2 object that contains the matrices M0, M1, M2, B0, B1 and B2,
that are necessary for computing the precision matrix of the Gaussian Matérn
field used for creating the SPDE approximation.
y0: A vector of equal length to the data (y_inla) used for running R-INLA,
such that y0\link{i} contains the value at the conditioning site for the same
time as when y_inla\link{i} was observed.
n: A vector of length equal to the number of unique conditioning sites used
for inference. Element nr. i of n contains the number of threshold exceedances
found at conditioning site nr. i.
init: Initial values for the subset of (log(λ), log(κ), log(ρ), log(σ), log(ρ_b), log(τ))
that will be estimated. See the documentation for the prior variable for more information.
priors: Priors for the six parameters of the model. This is a list of length 6,
with names "lambda", "kappa", "rho", "sigma", "rho_b" and "tau".
The list elements are vectors of
length 2 or 3. If the first element of a vector is 0, then we will not estimate
that parameter, but we will fix it equal to the value given in the second element
of the vector. If the first element is not 0, then the vector has length 3, and
the two remaining values are used for placing a prior on that parameter.
log(λ), log(κ) and log(ρ_b) are given Gaussian priors.
ρ and σ are given a joint PC prior, see ?inla.spde2.pcmatern for more information.
τ is given a gamma prior.
dist_to_s0: A vector of equal length to y_inla, such that dist_to_s0\link{i} contains
the distance from the location where y_inla\link{i} was observed, to the conditioning
site where y0\link{i} was observed.
dist_to_s0_from_mesh: A matrix of dimension (n x m), where m is the number of
nodes in the triangular mesh, and n is the number of unique conditioning sites.
Row nr. i of the matrix contains the distances from all mesh nodes to the
ith conditioning site. dist_to_s0_from_mesh can also be a vector, if only
one conditioning site is used.
C: NULL or a matrix that can be used to transform the prior.
theta_star: The mode of the likelihood. This is only used is C is non-null,
and it must be non-null if C is non-null.
debug: A boolean stating if R-INLA should print debug information or not.
}
