% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-generic-models.R
\name{spde_b_model_simulation}
\alias{spde_b_model_simulation}
\title{Function for defining a cgeneric model for Z_b.}
\usage{
spde_b_model_simulation(
  n,
  y0,
  spde,
  init,
  priors,
  dist_to_s0,
  is_fixed = rep(FALSE, 5),
  debug = FALSE
)
}
\description{
The input variables are:
n: A vector of length equal to the number of unique conditioning sites used
for inference. Element nr. i of n contains the number of threshold exceedances
found at conditioning site nr. i.
spde: A list with length equal to n, of inla.spde2 objects that each contains
the matrices M0, M1, M2, B0, B1 and B2,
which are necessary for computing the precision matrix of the Gaussian Matérn
field used for creating the SPDE approximation.
init: Initial values for the model parameters
priors: Priors for the model parameters. This is a list with one element per model parameter,
with names "sigma", "rho", "beta0", "lambda" and "kappa". Each element contains hyperparameters
for the prior of that specific model parameter.
dist_to_s0: A list of length equal to the n, where element nr. i contains the distances
from all the mesh nodes in spde\link{i} to conditioning site nr. i.
is_fixed: A vector of bools, stating whether each model parameter should be fixed to its
initial value or if it should be estimated.
debug: A boolean stating if R-INLA should print debug information or not.
}
