% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tikz_plot}
\alias{tikz_plot}
\title{Turn an R plot into a beautiful pdf made by LaTeX and TikZ,
using the tikzDevice package}
\usage{
tikz_plot(file, plot = NULL, expression = NULL, ...)
}
\description{
Turn an R plot into a beautiful pdf made by LaTeX and TikZ,
using the tikzDevice package
}
